+++ apps/storybook/.storybook/main.js
@@ -1,0 +1,16 @@
+module.exports = {
+  "stories": [
+    "../../../packages/*/src/**/*.stories.@(tsx)",
+  ],
+  "addons": ["@storybook/addon-links", "@storybook/addon-essentials", "@storybook/addon-interactions"],
+  "framework": {
+    name: "@storybook/react-vite",
+    options: {}
+  },
+  "features": {
+    "storyStoreV7": true
+  },
+  docs: {
+    autodocs: true
+  }
+};