+++ packages/layout/src/components/MainMenu.tsx
@@ -1,3 +1,4 @@
+import { useTranslation } from "@dzangolab/react-i18n";
 import { ResponsiveMenu } from "@dzangolab/react-ui";
 import React from "react";
 
@@ -11,22 +12,33 @@ interface Properties {
 
 const DEFAULT_ROUTES = [
   {
-    name: "Home",
+    name: "header.menu.home",
     route: "/",
   },
   {
-    name: "About",
+    name: "header.menu.about",
     route: "/about",
   },
 ];
 
 const MainMenu = (properties: Properties) => {
-  const { routes = DEFAULT_ROUTES, horizontal } = properties;
+  const { t } = useTranslation("app");
+
+  const { routes = DEFAULT_ROUTES, horizontal = false } = properties;
+
+  const getTranslatedRoutes = () => {
+    return routes.map(({ name, route }) => {
+      return {
+        name: t(name),
+        route,
+      };
+    });
+  };
 
   return (
     <ResponsiveMenu
       cssClass="main-menu"
-      routes={routes}
+      routes={getTranslatedRoutes()}
       horizontal={horizontal}
     />
   );