+++ packages/ui/src/Loading/index.ts
@@ -1,4 +1,0 @@
-import LoadingIcon from "./LoadingIcon";
-import LoadingPage from "./LoadingPage";
-
-export { LoadingIcon, LoadingPage };