+++ packages/user/src/index.ts
@@ -4,15 +4,23 @@ import AuthGoogleCallback from "./components/AuthGoogleCallback";
 import DropdownUserMenu from "./components/DropdownUserMenu";
 import UserMenu from "./components/UserMenu";
 import UserProvider, { userContext } from "./context/UserProvider";
+import UserEnabledAdminLayout from "./layouts/UserEnabledAdminLayout";
+import UserEnabledBasicLayout from "./layouts/UserEnabledBasicLayout";
 import superTokens from "./supertokens";
 import UserToastContainer from "./toastify";
-import { UserContextType } from "./types";
+import { DzangolabReactUserConfig, UserContextType } from "./types";
 import ChangePassword from "./views/ChangePassword";
 import ForgetPassword from "./views/ForgetPassword";
 import Login from "./views/Login";
 import Profile from "./views/Profile";
 import ResetPassword from "./views/ResetPassword";
 import Signup from "./views/Signup";
+
+declare module "@dzangolab/react-config" {
+  export interface AppConfig {
+    user: DzangolabReactUserConfig;
+  }
+}
 
 export {
   AuthGoogleCallback,
@@ -23,6 +31,8 @@ export {
   Profile,
   ResetPassword,
   Signup,
+  UserEnabledAdminLayout,
+  UserEnabledBasicLayout,
   UserMenu,
   UserProvider,
   UserToastContainer,