+++ packages/user/src/types/config.ts
@@ -1,0 +1,11 @@
+import type { RouteOverrides } from "./router";
+
+interface DzangolabReactUserConfig {
+  redirectTo: {
+    appName: string;
+    appURL: string;
+  };
+  routes?: RouteOverrides;
+}
+
+export type { DzangolabReactUserConfig };