+++ packages/user/src/types/index.ts
@@ -1,0 +1,9 @@
+import { DzangolabReactUserConfig } from "./config";
+import { AuthState, UserContextType, LoginCredentials } from "./types";
+
+export type {
+  AuthState,
+  DzangolabReactUserConfig,
+  LoginCredentials,
+  UserContextType,
+};