+++ packages/user/src/views/Login.tsx
@@ -7,6 +7,7 @@ import { toast } from "react-toastify";
 
 import GoogleLogin from "@/components/GoogleLogin";
 import LoginForm from "@/components/LoginForm";
+import RedirectionMessage from "@/components/RedirectionMessage";
 import login, { verifySession } from "@/supertokens/login";
 
 import { userContext } from "../context/UserProvider";
@@ -19,26 +20,62 @@ const Login = () => {
   const { t } = useTranslation("user");
   const { setUser } = useContext(userContext) as UserContextType;
   const [loading, setLoading] = useState<boolean>(false);
+  const [showRedirectionMessage, setShowRedirectionMessage] =
+    useState<boolean>(false);
+
   const appConfig = useContext(configContext);
 
   const handleSubmit = async (credentials: LoginCredentials) => {
     setLoading(true);
     const result = await login(credentials);
 
-    if (
-      result?.user &&
-      appConfig?.appContext &&
-      (await verifySession(appConfig.appContext))
-    ) {
-      setUser(result?.user);
-      toast.success(`${t("login.messages.success")}`);
+    if (result?.user) {
+      if (
+        appConfig?.appContext &&
+        (await verifySession(appConfig.appContext))
+      ) {
+        setUser(result?.user);
+        setShowRedirectionMessage(false);
+        toast.success(`${t("login.messages.success")}`);
+      } else {
+        setShowRedirectionMessage(true);
+      }
     }
     setLoading(false);
+  };
+
+  const getLinks = () => {
+    return (
+      <>
+        {appConfig?.user?.routes?.signup?.disabled ? null : (
+          <Link to={appConfig?.user?.routes?.signup?.path || "/signup"}>
+            {t("login.links.signup")}
+          </Link>
+        )}
+        {appConfig?.user?.routes?.forgetPassword?.disabled ? null : (
+          <Link
+            to={
+              appConfig?.user?.routes?.forgetPassword?.path ||
+              "/forget-password"
+            }
+          >
+            {t("login.links.forgotPassword")}
+          </Link>
+        )}
+      </>
+    );
   };
 
   return (
     <div className="login">
       <Page title={t("login.title")}>
+        {showRedirectionMessage ? (
+          <RedirectionMessage
+            appLink={appConfig?.user.redirectTo.appURL || ""}
+            appName={appConfig?.user.redirectTo.appName || ""}
+            hideRedirectionMessage={() => setShowRedirectionMessage(false)}
+          />
+        ) : null}
         <LoginForm handleSubmit={handleSubmit} loading={loading} />
         {appConfig?.supportedLoginProviders &&
         appConfig.supportedLoginProviders.includes("google") ? (
@@ -50,10 +87,7 @@ const Login = () => {
         ) : (
           <></>
         )}
-        <div className="links">
-          <Link to="/signup">{t("login.links.signup")}</Link>
-          <Link to="/forget-password">{t("login.links.forgotPassword")}</Link>
-        </div>
+        <div className="links">{getLinks()}</div>
       </Page>
     </div>
   );